pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "django-guestman"
version = "0.5.0"
description = "Customer management for Django"
readme = "README.md"
license = "MIT"
authors = [
    {name = "Pablo Valentini", email = "pablondrina@gmail.com"}
]
maintainers = [
    {name = "Pablo Valentini", email = "pablondrina@gmail.com"}
]
keywords = [
    "django",
    "customer",
    "crm",
    "contacts",
    "omnichannel",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 5.0",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.11"
dependencies = [
    "Django>=5.0,<6.0",
]

[project.optional-dependencies]
unfold = [
    "django-unfold>=0.49,<1.0",
]
admin = [
    "django-unfold>=0.49,<1.0",
]
dev = [
    "pytest>=7.0",
    "pytest-django>=4.5",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/pablondrina/django-guestman"
Repository = "https://github.com/pablondrina/django-guestman"
Changelog = "https://github.com/pablondrina/django-guestman/blob/main/CHANGELOG.md"
Issues = "https://github.com/pablondrina/django-guestman/issues"

[tool.setuptools]
packages = [
    "guestman",
    "guestman.adapters",
    "guestman.api",
    "guestman.models",
    "guestman.services",
    "guestman.signals",
    "guestman.protocols",
    "guestman.management",
    "guestman.management.commands",
    "guestman.migrations",
    "guestman.tests",
    "guestman.contrib",
    "guestman.contrib.admin_unfold",
    "guestman.contrib.identifiers",
    "guestman.contrib.identifiers.migrations",
    "guestman.contrib.insights",
    "guestman.contrib.insights.migrations",
    "guestman.contrib.preferences",
    "guestman.contrib.preferences.migrations",
    "guestman.contrib.customer_lookup",
    "guestman.contrib.manychat",
    "guestman.contrib.analytics",
]
include-package-data = true

[tool.setuptools.package-dir]
guestman = "."

[tool.setuptools.package-data]
guestman = [
    "py.typed",
]

[tool.ruff]
line-length = 120
target-version = "py311"
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008"]
exclude = ["migrations"]

[tool.ruff.isort]
known-first-party = ["guestman"]
known-third-party = ["django"]

[tool.pytest.ini_options]
python_files = ["test_*.py", "*_test.py"]
addopts = ["--strict-markers", "-ra"]

[tool.coverage.run]
source = ["guestman"]
omit = ["*/migrations/*", "*/tests/*"]
